Add review delete functionality and random_username session storage
This commit includes multiple updates. Firstly, it adds a function 'deleteReview' in the frontend 'ReviewsDisplay.vue' to allow review authors to delete their own reviews, with checks to prevent unauthenticated or wrong users from deleting reviews. An accompanying style for the review delete icon is also added.

Additionally, the authenticated user's random_username is saved and retrieved from local storage for user session continuity in 'userStore.js'. This is for correctly displaying the authorship of reviews.

Accompanying changes are made to the backend, including a new 'delete' endpoint in 'review.py' and inclusion of random_username in the response in 'auth.py'.

Finally, the commit modifies the server settings in 'settings.py' and switch to HTTP from HTTPS in 'axiosConfig.js'. The change in settings is to better integrate Django with the frontend and localhost. And the switch to HTTP is for local development convenience.