Remove print statement from auth.py
Removed an unnecessary print statement in the login function of the auth.py file. This change was made to clean up the console output and improve readability during debugging. This print statement was initially used for debugging the random_username and it is no longer needed since the functionality of random_username has been confirmed.