Move delete icon to the top right of each review
Moved the trash icon to the top right of each review if the current user is the author of the review. This change was made to improve visibility and accessibility of the delete button. Before, the delete icon was situated below the review, which was easily overlooked at and confusing due to the proximity of the like button.